исправлена ошибка. благодаря типу данных double можно решать задачу для больших чисел

+++ hseolymp/2016-17/demo/mars.cpp
@@ -5,14 +5,13 @@ using namespace std;
 
 int main()
 {
-	int n;
+	double n;
 	cin >> n;
 	int days = 0;
-	int factory = 1;
+	double factory = 1.0;
 
 	int without_fact = ceil((factory + sqrt(factory * factory + 8 * n * factory)) / (2 * factory));
 	int with_fact = ceil((2 * factory + sqrt(4 * factory * factory + 16 * n * factory)) / (4 * factory) + 1);
-
 	while (with_fact < without_fact)
 	{
 		factory *= 2;
@@ -20,7 +19,6 @@ int main()
 		without_fact = ceil((factory + sqrt(factory * factory + 8 * n * factory)) / (2 * factory));
 		with_fact = ceil((2 * factory + sqrt(4 * factory * factory + 16 * n * factory)) / (4 * factory) + 1);
 	}
-
 	days += ceil((factory + sqrt(factory * factory + 8 * n * factory)) / (2 * factory));
 	cout << days << endl;
 	return 0;
